+++ pyodide_lock/__init__.py
@@ -1,7 +1,8 @@
-from .spec import PyodideLockSpec
+from .spec import PackageSpec, PyodideLockSpec
 from .utils import parse_top_level_import_name
 
 __all__ = [
     "PyodideLockSpec",
+    "PackageSpec",
     "parse_top_level_import_name",
 ]