+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Adds `PackageSpec.from_wheel` for generating a package spec from a `.whl` file
+  [#18](https://github.com/pyodide/pyodide-lock/pull/18)
 - Adds `parse_top_level_import_name` for finding importable names in `.whl` files
   [#17](https://github.com/pyodide/pyodide-lock/pull/17)